Add event listeners, import wait-on-lib and include flapper.js linked from Google Drive
Using wait on lib package to import an external flapper.js file that is
hosted on Google Drive.

Added helper to main and as a result shifted main.html to a template.

Created the button to and tied listener to the button such that when
pressed, the jquery would run.